requirements.txt
bcrypt==4.0.1
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
Flask==2.2.2
Flask-SQLAlchemy==3.0.2
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
pycparser==2.21
requests==2.28.1
six==1.16.0
SQLAlchemy==1.4.42
urllib3==1.26.12
Werkzeug==2.2.2
google-auth==2.17.3
google-auth-oauthlib==0.4.6
google-auth-httplib2==0.1.0
google-api-python-client==2.24.0
openai==0.19.0
python-dotenv==1.0.0